refactor(commands): extract TransportInvoker for the transport-call wrapping layer
Move the transport-wrapping cluster (measureTransportCall, createProgressHandler,
ensureTransportSession, checkTransportSession, promptTransportSession,
setMode/setModel/getModel/cancelTransportSession, refreshSessionTransportAgentCommand,
plus isAbortError/inferTransportKind) verbatim into TransportInvoker. Router
composes one instance; ops factories + the two CRUD methods still resident
re-point to it. autoInstall/discoverPaths passed as late-binding lambdas so
__setAutoInstallForTest still reaches the invoker.

Zero behaviour change: 18 golden oracle fixtures byte-identical; mutation checks
T1/T2 reddened only their target scenarios; tsc clean.
